+++ src/js/_store/modules/categories.js
