+++ src/js/_store/modules/index.js
@@ -1,7 +1,7 @@
 import articles from './articles';
 import auth from './auth';
 import users from './users';
-import categories from './categoris';
+import categories from './categories';
 
 export {
   articles,
